+++ services/workflows-service/src/workflow-defintion/workflow-definition.service.ts
@@ -13,6 +13,7 @@ import { Injectable } from '@nestjs/common';
 import { Prisma, WorkflowDefinition } from '@prisma/client';
 import { merge } from 'lodash';
 import { upsertDemoEuKybFlow } from './demo-workflow/eu/create-demo-workflow-eu';
+import { upsertDemoUsKybFlow } from './demo-workflow/us/create-demo-workflow-us';
 
 @Injectable()
 export class WorkflowDefinitionService {
@@ -269,6 +270,10 @@ export class WorkflowDefinitionService {
 
   async createDemoWorkflowDefinitionEu({ customerId }: { customerId: string }) {
     return await upsertDemoEuKybFlow(customerId);
+  }
+
+  async createDemoWorkflowDefinitionUs({ customerId }: { customerId: string }) {
+    return await upsertDemoUsKybFlow(customerId);
   }
 
   async getByWorkflowRuntimeDataId(workflowRuntimeDataId: string, projectIds: TProjectIds) {