+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-core
 
+## 0.6.126
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.103
+
 ## 0.6.125
 
 ### Patch Changes