+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.126
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.126
+
 ## 0.6.125
 
 ### Patch Changes